requirements.txt
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
streamlit
python-dotenv
tiktoken